ioctl cmd arg is u_long
SCCS-vsn: 8.4

+++ sys/netinet/in.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)in.c	8.3 (Berkeley) 11/07/94
+ *	@(#)in.c	8.4 (Berkeley) 01/09/95
  */
 
 #include <sys/param.h>
@@ -158,7 +158,7 @@ extern	struct ifnet loif;
 /* ARGSUSED */
 in_control(so, cmd, data, ifp)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	register struct ifnet *ifp;
 {
